add Health Checker

+++ go-backend/agent/agent.go
@@ -8,7 +8,7 @@ import (
 type Agent struct {
 	Id          int      `json:"id"`
 	Status      string   `json:"status"` //running/missing/dead
-	Last_active int      `json:"last_active"`
+	Last_active int64      `json:"last_active"`
 	Goroutines  int      // Количество горутин, которые сейчас задействует агент
 	Stop        chan int // Канал для остановки работы агента(нужен, если агент завис или связь с ним потеряна)
 }
